Include middleware in apollo-client docs (#95)
* Include middleware in apollo-client docs

I wrote examples of writing basic middleware for use with createNetworkInterface.  I don't go into detail regarding how it would be used to manipulate headers and such, but that could also be added if needed.

* Add HTTP headers to example code

Changed the example code to show how you'd modify a request using middleware, instead of just showing a `console.log()`.  

An example is shown that takes a token from localStorage and passes it in a header.