Instructions for Redux dev tools support included (#179)
As requested by @stubailo. The code examples have been changed throughout the document. If a separate section for this topic is preferred over this update (i.e. using dev tools with "custom redux store") just let me know. I however think this is the best way because Redux is at the heart of the Apollo Client, and so are the Redux dev tools during the development process. Most developers are likely to use it this way if they require a customised Redux store.